# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IncidentDesk",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/investigations": {
      "post": {
        "summary": "Atomically create investigation and queue job",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "service": {
                    "type": "string"
                  },
                  "symptom": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 4000
                  },
                  "start": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "end": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "timezone": {
                    "type": "string"
                  },
                  "query": {
                    "additionalProperties": false,
                    "properties": {
                      "mode": {
                        "default": "anomalies",
                        "enum": [
                          "anomalies",
                          "search",
                          "all"
                        ],
                        "title": "Mode",
                        "type": "string"
                      },
                      "terms": {
                        "items": {
                          "type": "string"
                        },
                        "maxItems": 6,
                        "title": "Terms",
                        "type": "array"
                      },
                      "levels": {
                        "items": {
                          "enum": [
                            "ERROR",
                            "FATAL",
                            "SEVERE",
                            "FAILURE",
                            "WARN",
                            "WARNING",
                            "CRITICAL",
                            "ALERT",
                            "EMERG",
                            "INFO",
                            "DEBUG",
                            "TRACE",
                            "UNKNOWN"
                          ],
                          "type": "string"
                        },
                        "maxItems": 12,
                        "title": "Levels",
                        "type": "array"
                      }
                    },
                    "title": "LogQuery",
                    "type": "object"
                  }
                },
                "required": [
                  "service",
                  "symptom",
                  "start",
                  "end",
                  "timezone"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ]
      },
      "get": {
        "summary": "Latest 50 accessible investigations; offset query pagination",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        }
      }
    },
...
    "/api/v1/investigations/{id}/events": {
      "get": {
        "summary": "JSON events or SSE; Last-Event-ID cursor, reconnect after 25 seconds",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/internal/v1/claim": {
      "post": {
        "summary": "Atomic lease claim; 204 when no slot or work",
        "security": [
          {
            "WorkerKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "owner": {
                    "type": "string"
                  }
                },
                "required": [
                  "owner"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Protocol-Version",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "OIDC JWT"
      },
      "SourceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Source-Key"
      },
      "WorkerKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Worker-Key"
      }
    }
  }
}